refactor(*): remove clauseReducer to instead put clauses on contract in the the store
