+++ src/actions/contractActions.js
@@ -14,3 +14,15 @@ export const addToContractAction = uri => ({
   type: 'ADD_TO_CONTRACT',
   uri,
 });
+
+export const parseClauseSuccess = (clauseId, parseResult) => ({
+  type: 'PARSE_CLAUSE_SUCEEDED',
+  clauseId,
+  parseResult,
+});
+
+export const parseClauseError = (clauseId, error) => ({
+  type: 'PARSE_CLAUSE_ERROR',
+  clauseId,
+  error,
+});
