+++ src/reducers/contractReducer.js
@@ -19,6 +19,7 @@ const initialState = {
       }]
     }
   }),
+  clauses: {},
 };
 
 const reducer = (state = initialState, action) => {
@@ -28,6 +29,29 @@ const reducer = (state = initialState, action) => {
         ...state,
         markdown: action.markdown,
         slateValue: action.slateValue,
+      };
+    case 'PARSE_CLAUSE_SUCEEDED':
+      return {
+        ...state,
+        clauses: {
+          ...state.clauses,
+          [action.clauseId]: {
+            parseError: null,
+            parseResult: action.parseResult
+          }
+        }
+      };
+    case 'PARSE_CLAUSE_ERROR':
+      return {
+        ...state,
+        clauses: {
+          ...state.clauses,
+          [action.clauseId]: {
+            ...state[action.clauseId],
+            parseError: action.error,
+            parseResult: null, // would we rather have this be that last good result (if there is one)?
+          }
+        }
       };
     default:
       return state;
