+++ src/store.js
@@ -4,11 +4,12 @@ import logger from 'redux-logger';
 
 import appReducer from './reducers/appReducer';
 import templatesReducer from './reducers/templatesReducer';
-import modelReducer from './reducers/modelReducer';
-import logicReducer from './reducers/logicReducer';
-import sampleReducer from './reducers/sampleReducer';
-import clauseReducer from './reducers/clauseReducer';
+import modelReducer from './reducers/modelReducer'; // remove
+import logicReducer from './reducers/logicReducer'; // remove
+import sampleReducer from './reducers/sampleReducer'; // remove
 import contractReducer from './reducers/contractReducer';
+import clauseTemplatesReducer from './reducers/clauseTemplatesReducer';
+import contractTemplatesReducer from './reducers/contractTemplatesReducer';
 import rootSaga from './sagas/rootSaga';
 
 const sagaMiddleware = createSagaMiddleware();
@@ -20,12 +21,13 @@ if (process.env.NODE_ENV === 'development') {
 
 const rootReducer = combineReducers({
   appState: appReducer,
-  templatesState: templatesReducer,
-  modelState: modelReducer,
-  logicState: logicReducer,
-  sampleState: sampleReducer,
-  clauseState: clauseReducer,
+  templatesState: templatesReducer, // rename???
+  modelState: modelReducer, // remove
+  logicState: logicReducer, // remove
+  sampleState: sampleReducer, // remove
   contractState: contractReducer,
+  clauseTemplatesState: clauseTemplatesReducer,
+  contractTemplatesState: contractTemplatesReducer,
 });
 
 const store = createStore(
